exit with nonzero status code if lost images are found
Originally, the decision was made to still exit without an error even if
there were missing images in the source registry, because of two reasons:

(1) we wanted to continue to promote as many images as possible, even in
the face of errors (i.e., do not exit prematurely if errors are
encountered)

(2) if we are adding an image "foo" to the manifest, we should not error
out because of problems with some *other* image

In hindsight, (2) is misguided: the promoter manifest is the source of
truth *in its entirety*, so any errors that deviate from it must be
treated as a serious error. And (1) is not even a problem because we can
exit with an error by just remembering that there was indeed an error
somewhere along the promotion process (and proceeding to promote as many
images as we can regardless).

Because of this, we now exit with a nonzero error if any images are lost
from the source registry.