Manifest: deprecate top-level `src-registry` field
Instead, one of the registries in the `registries` list must contain the
line `src: true` to mark it as a source registry.

We retain a pointer to the source registry in the outer SyncContext
struct, for convenience. This has the added benefit of simplifying the
`mkPopReq` function as it no longer needs to have the Manifest passed
into it as the first argument.