use HTTPS GET for reading the state of the world
Replace the ReadImageNames/ReadDigestsAndTags combo with a recursive
walk that uses an HTTP API call to list all repo contents.

It's a nested tree structure, but we don't need to use a tree data
structure at all because the MasterInventory will store all of these
under the single "gcr.io/louhi-gke-k8s" registry entry (under a single
map); it's just that the image names will now include slashes in
them (for child repositories with images in them).

Here is what an example API call and reponse looks like from the CLI:

$ curl -u "oauth2accesstoken:$(gcloud auth print-access-token)" https://gcr.io/v2/louhi-gke-k8s/glbc/tags/list | jq
{
  "child": [
    "foo"
  ],
  "manifest": {
    "sha256:c3d310f4741b3642497da8826e0986db5e02afc9777a2b8e668c8e41034128c1":
  {
      "imageSizeBytes": "12875324",
      "layerId": "",
      "mediaType": "application/vnd.docker.distribution.manifest.v2+json",
      "tag": [
        "deleteme"
      ],
      "timeCreatedMs": "1501774217070",
      "timeUploadedMs": "1552917295327"
    }
  },
  "name": "louhi-gke-k8s/glbc",
  "tags": [
    "deleteme"
  ]
}

NOTE: There is now a slight change of WaitGroup (semaphore) semantics.

Previously, the semaphore (wg.Add() and wg.Done()) was used to only
count the number of workers (actual number of goroutines). We would know
the finite number of jobs to execute, and after creating all the jobs,
we would close the "reqs" channel, at which time all workers would die
and decrement the semaphore.

Now the semaphore tracks the number of pending requests that must be
performed). I.e., every request that gets created increments the
semaphore, and every request that gets processed decrements it. This
sort of abuses the meaning of the WaitGroup counter, as the official
docs say that it should equal the number of goroutines [1]:

	The main goroutine calls Add to set the number of goroutines to wait
	for.

The alternative to this is to just use a global semaphore with locks to
increment decrement it, and then to just poll against this semaphore
from the outside (and to kill all goroutines when it reached 0). While
this is doable, the code becomes more verbose; because WaitGroup.Add()
is already an int and alread atomic, we just use reuse it here with a
slight change in meaning.

[1]: https://golang.org/pkg/sync/#WaitGroup.Add