vendor all dependencies
This adds gopkg.in/yaml.v2 and github.com/google/go-containerregistry
toplevel dependencies (go-containerregistry brings with it some others,
as it is not self-contained). go-containerregistry is needed because we
will use the data structures it defines for GCR API responses in the
near future when we leverage it instead of shelling out to `gcloud`.

Vendoring is really ugly, but it's how things are done in the Go world.