convert -gen-images flag to -snapshot
The changes here are in preparation for the e2e test. The new -snapshot
flag (which deprecates the -gen-images flag) allows us to snapshot
repository contents (the same contents always output the same YAML,
because we sort the hashmap items). This ability to snapshot
repositories gives us a "fingerprint" of a GCR, so that we can use
predefined fingerprints to check against a live repository (the
test-only GCRs) during tests.

As for the change here itself (in comparison to the old GenImages()
function), instead of requiring the user to specify a specialized
Manifest file, we have the user specify a single repository as a
command-line argument (along with an optional service account flag if
the repository is private). Then, we manually construct a Manifest
internally before running the ReadRepository logic.