add -manifest-dir flag: allow reading manifests recursively
Also, rework how promotions are done. We use a new concept called a
"PromotionEdge" to encode the promotions we want to perform. This
simplifies the logic for mkPopulateRequestsForPromotion (now
deprecated); compare with mkPopulateRequestsForPromotionEdges.

Note: The deprecated function mkPopulateRequestsForPromotion was not
deleted in this commit to make it easier to compare the functions (and
to make the diff simple).