e2e: refactor to use promoter Manifest files in the filesystem
Although this breaks the "have one place for everything" rule that is
generally preferred when writing test cases, it is better because we no
longer have to write temporary Manifest files. Additionally we have less
of a dependency on the Manifest (single) type, which will be handy when
we write a test case for recursive Manifest directories (where there can
be 2 or more Manifests).