ReadRegistries: adjust sc.Inv layout to match Manifest
Before, we stored keys into sc.Inv by relying on
GetTokenKeyDomainRepoPath. Now, we just split along the registry path
defined in the Manifest.

E.g., before if we had a Manifest that stated

    registries:
    - name: gcr.io/a/b/c
    ...
    images:
    - name: d
    ...

and read in

    "gcr.io/a/b/c/d"

as part of ReadRegistries(), we would store it as

    sc.Inv["gcr.io/a"]["b/c/d"]...

but now we store it as

    sc.Inv["gcr.io/a/b/c"]["d"]...

to match the delineations in the Manifest.