e2e: use download_pkgs() instead of dpkg_list() for GNU make
This is because using dpkg_list of Debian Jessie snapshot results in
taking 10 minutes+ to join the Docker layers together (the Ubuntu 16.04
base image and Debian Jessie probably have a large delta in their
filesytem layouts, resulting in a very long time to create the image via
rules_docker's join_layers.py).

Using download_pkgs() is more correct, because we are using the base
image's apt sources to pull the right version of GNU make.