Support transparent compression on GCS
We attempt to compress files as part of the upload, and upload a
compressed form if the compression was worthwhile.  We store metadata
in the file attributes, particularly because the md5 is now the md5 of
the compressed file, so need no longer match the source md5.  We
instead store the sha256, and match this against the manifest.