gcloud: add ActivateServiceAccounts
This is a port of the account activation logic that is in the
to-be-deprecated util/multirun.sh. The cip binary will use it so that
multirun.sh can be removed.

At the same time, move activateServiceAccount under the gcloud package
because that's where it should live (there are other gcloud-specific
things in inventory.go but that can happen later).