SplitByKnownRegistries: sort names for determinism
Otherwise, the order of looping in SplitByKnownRegistries is
non-deterministic and in the case where there are 2 known registries
with a shared leading prefix string, we might select the wrong one.

E.g., if we have "gcr.io/foo" and "gcr.io/foo-bar", we might end up
splitting the string "gcr.io/foo-bar/qux" with the first one, giving us
"gcr.io/foo" and "-bar/qux" (incorrect). This patch prevents this
scenario.