sort SyncContext.RegistryContexts for determinism
This is a follow-up to 0b2469b15ab754e7ad80fff2c362467bcdb4477d. The
problem with sorting things in SplitByKnownRegistries is that that
function was running inside a goroutine, modifying the global
SyncContext structure without locking it first. Sorting the registries
in MakeSyncContext makes it so that we don't have to bother adding a
lock.