detect and prevent duplicate PromotionEdges with contrasting intent
It could be the case that two edges attempt to promote different
images (different digests) to the same destination endpoint. This
changes enables this detection.

We also add tests to detect cases where 2 manifests with different
digests are being promoted to the same destination image name by way of
a rename operation. Even though renames are not used in practice, it's
still a possibility we have to check for.