filter promotionEdges to remove what's been promoted
This change should have been in
2a09813ffe2f9ba84cf0530077bbf696c2151e18. In that commit the logic to
compare PromotionEdges against the state of the world (reading the
source/destination registries) was split out from
GetPromotionEdgesFromManifests() to a new function named
FilterPromotionEdges(), but we forgot to actually call it to perform the
filtering.

As this is an optimization step, it explains why the e2e tests for #132
passed. Even so, without this step we don't get any of the interesting
insights as to what images are missing, what are not, etc. so the
omission of calling this function in the earlier PR was a problem.