add -manifest-based-snapshot-of flag
This flag can condense all of the manifest information from a
-manifest-dir, and output a complete inventory of all images we would
like to promote (filtered by the given destination registry name). This
is useful if, for example, we would like to know the complete catalogue
of all images that we would like to "exist" at a destination registry,
using the promoter manifests alone as the source of truth.

In a sense, this flag can be thought of as a sort of "manifest"
normalizer that can create a new manifest out of a bunch of smaller
manifest files (at least the `images:` portion of it).

This comes in handy in case we want to cross-reference the output of
-snapshot against a known good set of promoter manifests that are solely
responsible for adding new images to the snapshotted repo, for purposes
of auditing.