add -thin-manifest-dir flag
This is just like -manifest-dir, but instead of expecting promoter
manifests that have both

    registries: ...

and

    images: ...

information, it only expects so-called thin manifests that look like

    registries: ...
    imagesPath: <path to images.yaml>

. These thin manifests are designed to be owned by a select few, and
they will refer to images.yaml files that reside in other folder(s) with
less restrictive ACLs. The point is to force PRs to only change the
images.yaml portions of manifests, and *NEVER* the 'registries:' part,
which contains important metadata that should rarely change.