add basic auditing mechanism
This adds a new option to the promoter ("cip" binary) to let it listen
to HTTP requests that contain specially-crafted JSON payloads (from GCP
Cloud Pub/Sub). The idea is to get GCR to emit Pub/Sub events, and then
qualify every Pub/Sub message, by checking it with the state of the
promoter manifests.