disallow tag moves
Because tags should always be immutable, moving tags are no longer
supported. We leave the code around detecting such events in tact, in
case we want to revert this behavior in the future.

Note that this means if the destination registries get into a bad state
because of a system independent of the promoter, the promoter can no
longer be used to rectify the situation.