use Stackdriver-friendly logging and error reporting
Use the GCP Go client libs to perform logging and error reporting.
Specifically, we use the errorreporting API when recovering from panics
--- this lets us use the nice Error Reporting dashboard and alerting web
UI on GCP. Also, we use the logging API for logging things in the
Audit() function; among other things this prevents the logs from being
split up (which is the default behavior when we simply log things with
klog).