thin manifests: deprecate free-form ImagesPath field
We no longer allow an arbitrary location for the ImagesPath field (this
field is deprecated altogether and is not used). Instead, the following
scheme is used:

    foo
    ├── images
    │   ├── a
    │   │   └── images.yaml
    │   ├── b
    │   │   └── images.yaml
    │   ├── c
    │   │   └── images.yaml
    │   └── d
    │       └── images.yaml
    └── manifests
        ├── a
        │   └── promoter-manifest.yaml
        ├── b
        │   └── promoter-manifest.yaml
        ├── c
        │   └── promoter-manifest.yaml
        └── d
            └── promoter-manifest.yaml

where "foo" is the argument given to -thin-manifest-dir.

This change only enforces the rule that for a given subproject's
promoter-manifest.yaml file, that it must look for an images.yaml file
at: "./../../images/<SUBPROJECT_NAME>/images.yaml".

The next thing to do is to enforce the tree structure as outlined above
before parsing even begins (for validation).