address review comments
In particular, we don't use reflect.DeepEqual because we are even more
strict about which folders we traverse over when discovering
"promoter-manifest.yaml" files.

This new level of strictness makes the previously-invalid
"invalid/malformed-directory-tree-structure-bad-prefix" test case a
valid one (because we no longer traverse the troublesome
"manifest" (note the lack of a trailing "s") folder) --- and so that
test has been renamed to
"valid/malformed-directory-tree-structure-bad-prefix-is-ignored".