add auditor testing entrypoint
For now, this test doesn't really do anything. It is here mainly to
allow for a Prow check to be written that will use this script as the
entrypoint.

Later when the actual auditor E2E test is ready, we can create a PR for
that (at which point in time the Prow check above will activate and run
the full test suite for us).