convert LoggingFacility from a struct to an interface
This makes it so that the fake/real implementations don't have to share
anything. Whereas before, the implementations had to call the New()
constructor to create a LoggingFacility struct type, with this change
they can each have their own struct which just satisfies the
LoggingFacility interface.

The change is subtle but leads to lighter shared code in
logclient/types.go.