Audit: fix bug around finding the subproject registry
This part of the code deals with handling child image manifests whose
digests are NOT recorded in the promoter manifests [1]. Such use cases
are perfectly valid because it may be the case that users only want to
track the parent digest, not all of the component child digests as well.
So in the event that a child image's digest is received by the auditor,
it must only reject if the corresponding parent digest cannot be found
in the promoter manifests.

The fix involves using the improved Contains() method (which now has
much higher granularity about how close of a match the GCR payload for
the child image had with a promoter manifest) instead of using a very
crude suffix matching algorithm.

[1]: https://github.com/kubernetes-sigs/k8s-container-image-promoter/issues/165