TestAudit: add test case for verified manifest (simplest case)
This case covers the situation when a Pub/Sub payload has both the
digest and tag of an image and it matches exactly as in a promoter
manifest.

This is the simplest case of a verifiable match (direct match, as
opposed to a child manifest which might involve an indirect match).