use more granular fields for matching
This increases code readability, because we match on smaller
pieces (digest, tag) directly. It also makes the code faster because we
no longer have to construct FQINs and PQINs.

Also, we match more strictly for FlagPathMatched. Instead of just
checking for a prefix, we match on the entire path for an exact match.
This should make it more robust for cases where we have nested projects.