SplitByKnownRegistries: ensure clean split
We allowed cutting into the middle of an image. This resulted in repo
reads being mis-written into the wrong key, when populating the
SyncContext.Inv, in ReadRegistries. For example, the digests for

    us.gcr.io/k8s-artifacts-prod/metrics-server-amd64

were incorrectly assigned to the

    us.gcr.io/k8s-artifacts-prod/metrics-server

repo, under the image name "-amd64". This resulted in the promoter
thinking that there were no digests under
us.gcr.io/k8s-artifacts-prod/metrics-server-amd64 at all, resulting in
continued promotions to there. This has been fixed.

Also, this adds a unit test to ensure we don't suffer from a regression.