ToYAML: allow customizing marshaling style
The ToYAML() method is used for the RegInvImage type to generate
snapshots of registries, but it is now also used for modifying the
images section of thin promoter manifests. In order to make things more
sensible for mechanical diffing, the "SplitTagsOverMultipleLines" is
introduced. By default this is false, meaning that tags are always
listed on a single textual line. This has 2 advantages:

(1) diffs are easier to read, as a single line addition is responsible
for adding a single digest (regardless of the number of tags associated
with it), and

(2) it discourages tagging the same digest more than one time with a
different tag.