auditor: use shorter ExponentialBackoff for Cloud Run
Cloud Run has a 15 minute quota (or time limit) for requests to finish
"running". Otherwise the request is "terminated" [1]. Because we don't
want our requests to be rudely interrupted by Cloud Run, we use a
shorter (~13 minute) timeout, named BackoffCloudRun.

This in itself doesn't get us much, but it does take us out of
poorly-documented behavior (for example, it is unclear what exactly
happens to the container when the request is terminated abruptly by
Cloud Run).

[1]: https://github.com/kubernetes/k8s.io/pull/955#issuecomment-645794886