moved WaitGroup drain to after request results are processed
In order to fix the current bug in ExecRequests (Issue #237),
we can simply move the WaitGroup draining logic to after the
request results are processed. So from now on, the WaitGroup
is drained at the end of the ProcessRequest function if and
only if the request results channel is not being filled. Also,
for the logic to still work, the request results channel needs
to be filled at the end of the function.