e2e: update Image digests
The update to the new bazel build rules in
ca00a595859b0ee9b4729d6a430bda0b0bd16fad results in changed digests.
This commit updates the digests.

It's unfortunate that we have to manually do this (it would be far
better to somehow keep the same digests in the code), but it happens
rarely enough that it's probably not worth fixing.

It's also worth noting that the workaround of updating digests is
actually not that difficult. We just need to run some perl replacement
commands because the digests are UUIDs, like this:

    perl -pi -e 's/OLD_DIGEST/NEW_DIGEST/g' **/*.yaml