fix cip-postsubmit-push-to-staging error on tagged commits
Tagged commits result in the PROW_GIT_TAG resulting in the truncated
form vYYYYMMDD-tag instead of vYYYYMMDD-tag-n-ghash. This breaks the old
strict regex check and aborts the cip-postsubmit-push-to-staging Prow
job (which containerizes the latest master branch contents into a new
image and pushes it to gcr.io/k8s-staging-artifact-promoter).

Here we relax the strict regex check, and use fallback values for
git_commit and git_desc.