inventory_test.go: deprecate use of checkEqual and checkError
Instead we use the `require' package from the `testify' Golang library.
Previously, we used the pattern of first checking for equality with
checkEqual, followed by asserting that this checking operation itself
did not fail with checkError.

The `require.Equal()' function alone deprecates this pattern. The only
difference is that the ordering of (got, expected) is reversed
with (expected, actual), to align with the upstream documentation about
the usage of require.Equal(). The failure to observe this usage is part
of why https://github.com/kubernetes/release/pull/1772 introduced a
regression in these tests.