exit with error for tag moves
After we generate all PromotionEdges, we can determine if the intent in
these edges would result in a tag move (by comparing the information in
the PromotionEdge to the state of GCR). If the intent to move a tag is
detected, exit with an error.

This addresses #375. Although that bug only talks about dry runs, we
enforce this behavior for non-dry-runs as well (production runs),
because tag moves are not supported anyway expose invalid intent in the
promoter manifest YAMLs.