Dockerfile: Use an explicit go version (go1.16.7) for image building
Using a `latest` tag in general, but specifically for Golang-based
images will eventually lead to unexpected variance in our image builds.

Here we pull an explicit go version and add an entry for it in
dependencies.yaml.

This commit also adds a dependency entry for the go major version in
go.mod, to ensure it stays in sync with image builds.

Signed-off-by: Stephen Augustus <foo@auggie.dev>