filepromoter: Build unauthenticated client for source filestores
Source filestores are (supposed to be) world-readable, so there is
no need to require an authenticated client for them.

This is a workaround for the fact that service accounts are stored
as part of `filestore`, when ergonomically, they should probably be
at the top-level of a root promotion manifest. To be fixed in a
future commit.

Signed-off-by: Stephen Augustus <foo@auggie.dev>