cip subcommand: Drop legacy code in favor of new pkg
This commit drops all the legacy to code in the `kpromo cip`
subcommand and replacess it to 4 simple calls to the new
methods in the promoter package.

At this point all functionality should be reimplemented in
the new package and its implementation with the same algorithms
and calls to the rest of the packages.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>