Print error when no manifests found in tags list
When the sync contexts tries to pull the list of tags, it
depends on getting a list of manifests from there to traverse
the registry. If it does not get one, it will continue trying
but it will fail later on.

This commit modifies the request to print an error if no manifests
are found.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>