Add GetIdentityToken() function to the promoter implementation
This commit adds a new function GetIdentityToken to the promoter
implementation. This function allows the current workload identity
to impersonate a GCP service account and obtain an identity token from
it.

In order for the tests to work, the workload id that runs the test
has to have the needed permissions to impersonate the test SA.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@uservers.net>