Change signature annotation to version
Before this commit, we used to annotate the image signatures with
a record of the mirrors where we promoted the images. We need
an anotation in order to force cosign to add a signature to images
which are already promoted.

BNut now we are promoting to dozens of registries and it has made
the original annotation cumbersome and bloated.

This commit changes the annotation to record the promoter version
instead.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>